crawler/requirements.txt
requests
beautifulsoup4
cassandra-driver
flask
